🔬 fix: the placement spike asks the op whether it applied, not the writer (#1073)
Problem: since the placement gate moved into `toggleCell`, a refused placement
comes back as the INPUT model — and the input is the user's current model, which
serializes perfectly well. The #1058 spike inferred acceptance from a non-null
serialize, so every refusal was counted as an acceptance and the metric stopped
distinguishing the two at all. Re-run unmodified it reported `A = 100.0%` in every
cell of its table, and the refused clicks went on to reach the playback oracle as
edits and fail it — 9,082 corrupt asks against the 86 the committed artifact
records. Nothing failed, because the probe is a `.spec.ts` that no gate collects.

The hypothesis arm had the same hole in a worse form: it composes a cross-lane
clamp on top of the toggle, and that clamp cannot decline. It therefore clamped an
UNMODIFIED model, serialized it fine, and scored a rescue it never performed —
while the op it was built on refuses exactly the placements the arm exists to test.

Fix: ask the op for its own verdict, `op(x) !== x`. That is the signal the
notation op family already defines and the one every `can*` predicate is derived
from, so it cannot drift away from the op again. Applied at all five sites in the
probe. The hypothesis arm is rebuilt on the ungated toggle instead, which is the
pre-change shape the admissibility gate already keeps as its control arm — moved
into a shared module now that a second caller needs it, rather than copied.

Verified by reproduction: every figure in the committed artifact comes back, none
of them fed to the fix. A at all four depths (79.1 / 68.8 / 62.1 / 40.9), L (98.0 /
97.6 / 97.0 / 95.7), corrupt asks/units (3/2, 5/2, 9/2, 86/3), the base arm (11,633
asks, 85.0%, 1,717 of 1,748 declines by the same mechanism) and the hypothesis arm
(15,056 accepted, 14,778 local, 13,763 plays, 1,293 play-differs). The base arm's
1,748 declines of 11,633 asks independently match the figure recorded for the
element path in `place.ts`. The decision the artifact was taken on still holds.

Gates: parity-corpus 369/28, app 975/73, tsc 1 (pre-existing, untouched file).
